.github/update_localization_contributors.requirements.txt
crowdin-api-client
requests
